requirements.txt
customtkinter
PyGithub
requests
pyyaml
Jinja2
WeasyPrint
